kernels: repeat_interleave's per-element loop bound is 32-bit (Kokoro's next wall)
Once weight_norm was fixed, Kokoro advanced to aten.repeat_interleave and was
refused: "scf.for with 64-bit loop bounds: the induction lowering would not
terminate (hang). Use 32-bit bounds." `repeats` arrives i64 by torch default,
so `range(0, repeats, BLOCK)` has a 64-bit bound.

The refusal names its own remedy. `repeats` is cast to i32 for the loop:
safe -- a per-element repeat of 2**31 needs an 8 GB output for one element,
which the allocator refuses long before this loop -- and numerically inert on
CUDA (same iterations, narrower type). Addressing stays wide: `out_offset =
cumsum - repeats` promotes back through cumsum.

Verified on-device against torch.repeat_interleave: the index map is
bit-identical. The reduce family (swin2SR l2_norm, Kokoro weight_norm) is now
closed on both paths; this is the independent wall behind it, and it is our
kernel, not a lowerer refusal.

+++ src/neurobrix/kernels/ops/repeat_interleave.py
@@ -31,7 +31,14 @@ def repeat_interleave_tensor_kernel(
     pid = tl.program_id(0)
     mask = pid < size
     cumsum = tl.load(cumsum_ptr + pid, mask, other=0)
-    repeats = tl.load(repeats_ptr + pid, mask, other=0)
+    # The loop bound below must be 32-bit: the Metal induction lowering does
+    # not terminate on a 64-bit `scf.for` bound (it refuses rather than hang).
+    # `repeats` arrives i64 by torch default. The cast is safe -- a per-element
+    # repeat count of 2**31 would need an 8 GB output for ONE element, which
+    # the allocator refuses long before this loop -- and numerically inert on
+    # CUDA: same iterations, narrower type. Addressing stays wide: `out_offset`
+    # promotes back through `cumsum`.
+    repeats = tl.load(repeats_ptr + pid, mask, other=0).to(tl.int32)
     out_offset = cumsum - repeats
 
     base_ptr = out_ptr + out_offset
